Fixed defined presence of 0 being treated as actual presence.
This caused a new game to encounter pirates in Hakoi if a game that
had the pirates in Hakoi diff applied was loaded prior to starting
the new game. The reason for this is that when the unidiff system
removes presences, it doesn't remove the indicated presences number;
instead, it just decreases it. So pirates had a defined presence of
0 in Hakoi. Due to the way spawning works, this simply meant that
the spawner spawned pirates at the rate defined to be the minimum
spawn rate.